deleted BST file, added into names.py, reconfigured code, runtime less than one second.
