+++ names/names.py
@@ -1,5 +1,4 @@
 import time
-from binary_search_tree import BSTNode
 
 start_time = time.time()
 
@@ -14,19 +13,48 @@ f.close()
 duplicates = []  # Return the list of duplicates in this data structure
 
 # Replace the nested for loops below with your improvements
-for name_1 in names_1:
-    for name_2 in names_2:
-        if name_1 == name_2:
-            duplicates.append(name_1)
+# for name_1 in names_1:
+#     for name_2 in names_2:
+#         if name_1 == name_2:
+#             duplicates.append(name_1)
 
-bst = BSTNode("")
+class BSTNode:
+    def __init__(self, value):
+        self.value = value
+        self.left = None
+        self.right = None
 
-for name in names_1:
-    bst.inster(name)
+    def insert(self, value):
+        if value < self.value:
+            if not self.left:
+                self.left = BSTNode(value)
+            else:
+                self.left.insert(value)
+        elif value >= self.value:
+            if not self.right:
+                self.right = BSTNode(value)
+            else:
+                self.right.insert(value)
 
-for name in names_2:
-    if bst.contains(name):
-        duplicates.append(name)
+    def contains(self, target):
+        if target == self.value:
+            return True
+        elif target < self.value:
+            return self.left.contains(target) if self.left else False
+        else:
+            return self.right.contains(target) if self.right else False
+
+
+# runtime: 0.09576702117919922 seconds
+def bst():
+    every_name = BSTNode('Names')
+    for name in names_1:
+        every_name.insert(name)
+    for name in names_2:
+        if every_name.contains(name):
+            duplicates.append(name)
+
+bst()
 
 end_time = time.time()
 print (f"{len(duplicates)} duplicates:\n\n{', '.join(duplicates)}\n\n")
